Refactor most of stylesheet. Add new pages
Stylesheet uses variables for colors now so everything is a lot more consistent. I also cleaned up most of the doubles I found and redid the navbar so it's consistent on all pages. I added a section for the Unreal Jet project and attributions (just an overview page. No data yet)